Enabled only writing logic "0" on IO port, and adapted IO port names to FSM and logic i2c_ctrl.sv module
The names of the IN / OUT ports are adapted to the logic of the i2c_ctrl.sv module. Two new OUT ports scl_oe and sda_oe have been created, which put the I/O into high-impedance 3-state mode when
I/O is used for read (input) operation. In the write operation these signals set logic "0" on the I/O port, and leave the I/O port in 3-state when it write logic "1". In this way,  avoids the collision of simultaneous writing of multiple devices on I2C lines. Signals sda_do and scl_do now use only for internal logic instead for OUT ports.